Got year marker to move with cursor
Added styling to the year marker, and got it to move position based on the user's cursor (this code is in the canvas.js file). Generated some sprites based on dots.svg texture, and tried to get them to dynamically generate based on the number of movies but couldn't figure out how to access the variable from the app.js file. Tried to get the genre filter working, but not quite sure how to change <li ng-repeat="movie in movies"> to something like <li ng-repeat="movie in selectedGenre">