Moved canvas.js code into app.js file
I've moved all of the pixi code from canvas.js into the app.js file, so we could access the $scope.movies data. I haven't deleted the canvas.js file though (just commented it out in the html). I was able to generate a dot sprite for each movie, but haven't worked out how to display it according to the year yet.