Set delay on showing hover event
Set the delay on opening the hover event popup, but this seems to cause some issues with the loading time on images. Tried to make a pulse animation for the glow, but wasn't able to get it to work. Copied code from script.js file into app.js; deleted script.js.