Added animations to pop up div, fixed trailer click
Added animations when pop up div opens, but haven't been able to get the animation to occur when closed.

Because the height of the pop up div is set to auto, you can't use css transitions to animate it smoothly. To work around this, I've added in max-height: 500px, and overflow:scroll, in case there are any pop up divs larger than 500px (such as Resident Evil: Retribution). This makes the animation smooth, but has some side effects (like the image sometimes taking a while to load, or the actors fading in slightly after the rest of the content). I'll keep looking for a better fix, but this seems to be going ok.

Also fixed the link to the trailer. The detail-overview padding was on top of  the <a> tag, which is why it couldn't be clicked.